Update python
To check if a given string is numeric or not using Python regex (regular expressions), we can utilize the re.search() method. re.search() method either returns None (if the pattern doesn’t match) or a re.MatchObject contains information about the matching part of the string. This method stops after the first match, so this is best suited for testing a regular expression more than extracting data.